% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_basic_metadata}
\alias{add_basic_metadata}
\title{Add Basic Metadata}
\usage{
add_basic_metadata(
  metadata,
  input_tumor_type,
  input_tumor_details,
  input_disease,
  input_primary_site
)
}
\arguments{
\item{metadata}{A data.table containing metadata.}

\item{tumor_type}{The type of tumor.}

\item{tumor_details}{Details about the tumor.}

\item{disease}{The disease associated with the sample.}

\item{primary_site}{The primary site of the tumor.}
}
\value{
Updated metadata with basic information added.
}
\description{
Adds basic metadata information such as tumor type, tumor details, disease, and primary site.
}
