% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_fga}
\alias{add_fga}
\title{Add Fraction Genome Altered (FGA)}
\usage{
add_fga(metadata, jabba_gg = NULL, seqnames = c(1:22))
}
\arguments{
\item{metadata}{A data.table containing metadata}

\item{jabba_gg}{Path to JaBbA graph RDS file}

\item{seqnames}{Vector of sequence names to consider}
}
\value{
Updated metadata with LOH metrics
}
\description{
Add Fraction Genome Altered (FGA)
}
