% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_genome_length}
\alias{add_genome_length}
\title{Add Total Genome Length}
\usage{
add_genome_length(
  metadata,
  jabba_gg = NULL,
  seqnames_genome_width = c(1:22, "X", "Y")
)
}
\arguments{
\item{metadata}{A data.table containing metadata}

\item{jabba_gg}{Path to JaBbA graph RDS file}

\item{seqnames_genome_width}{Vector of sequence names to include in genome length}
}
\value{
Updated metadata with genome length information
}
\description{
Adds total genome length from JaBbA graph
}
