% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_het_pileups_parameters}
\alias{add_het_pileups_parameters}
\title{Add Heterozygous Pileups Parameters}
\usage{
add_het_pileups_parameters(metadata, het_pileups)
}
\arguments{
\item{metadata}{A data.table containing metadata}

\item{het_pileups_wgs}{Path to heterozygous pileups WGS data}
}
\value{
Updated metadata with heterozygous pileups parameters
}
\description{
Adds slope and intercept parameters calculated from heterozygous pileups
}
