% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_hrd_scores}
\alias{add_hrd_scores}
\title{Add HRD Scores}
\usage{
add_hrd_scores(metadata, hrdetect, onenesstwoness)
}
\arguments{
\item{metadata}{A data.table containing metadata.}

\item{hrdetect}{HRDetect scores.}

\item{onenesstwoness}{Oneness and twoness scores.}
}
\value{
Updated metadata with HRD scores added.
}
\description{
Adds homologous recombination deficiency (HRD) scores to the metadata.
}
