% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_loh}
\alias{add_loh}
\title{Add Loss of Heterozygosity Metrics}
\usage{
add_loh(metadata, jabba_gg = NULL, seqnames_loh = c(1:22))
}
\arguments{
\item{metadata}{A data.table containing metadata}

\item{jabba_gg}{Path to JaBbA graph RDS file}

\item{seqnames_loh}{Vector of sequence names to consider for LOH}
}
\value{
Updated metadata with LOH metrics
}
\description{
Adds LOH metrics from JaBbA graph
}
