% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_msisensor_score}
\alias{add_msisensor_score}
\title{Add MSIsensor Score}
\usage{
add_msisensor_score(metadata, msisensorpro)
}
\arguments{
\item{metadata}{A data.table containing metadata.}

\item{msisensor_pro}{Path to MSIsensor profile file.}
}
\value{
Updated metadata with MSIsensor score added.
}
\description{
Adds MSIsensor score to the metadata.
}
