% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_patient_to_distributions}
\alias{add_patient_to_distributions}
\title{add_patient_to_distributions}
\usage{
add_patient_to_distributions(
  metadata_json_path,
  common_dir,
  cores = 1,
  overwrite = FALSE,
  write_to_json = FALSE
)
}
\arguments{
\item{metadata_json_path}{Path to the new patient's metadata.json file.}

\item{common_dir}{Path to the common directory where the distributions are stored.}

\item{cores}{Number of cores to use for parallel processing.}

\item{overwrite}{Boolean flag to indicate whether to overwrite existing files.}
}
\value{
Updated distributions.
}
\description{
Adds a new patient to the existing distributions and writes the updated distributions to JSON files.
}
\author{
Shihab Dider
}
