% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_purity_ploidy}
\alias{add_purity_ploidy}
\title{Add Purity and Ploidy Metrics}
\usage{
add_purity_ploidy(metadata, jabba_gg = NULL, tumor_coverage = NULL)
}
\arguments{
\item{metadata}{A data.table containing metadata}

\item{jabba_gg}{Path to JaBbA graph RDS file}
}
\value{
Updated metadata with purity and ploidy information
}
\description{
Adds purity and ploidy information from JaBbA graph
}
