% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_sex_information}
\alias{add_sex_information}
\title{Add Sex Information}
\usage{
add_sex_information(
  metadata,
  input_inferred_sex = NULL,
  jabba_gg = NULL,
  tumor_coverage = NULL
)
}
\arguments{
\item{metadata}{A data.table containing metadata.}

\item{input_inferred_sex}{The inferred sex of the sample.}

\item{jabba_gg}{A file path to the jabba_gg data.}

\item{tumor_coverage}{Coverage data for the tumor.}
}
\value{
Updated metadata with inferred sex information added.
}
\description{
Adds inferred sex information to the metadata based on various inputs.
}
