% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_signature_cosine_similarity}
\alias{add_signature_cosine_similarity}
\title{Signature Cosine Similarity}
\usage{
add_signature_cosine_similarity(
  metadata,
  channel_counts = NULL,
  decomposed_signature_matrix = NULL,
  signature_reference_matrix = NULL,
  field_name
)
}
\arguments{
\item{metadata}{A data.table containing metadata.}

\item{channel_counts}{Channel mutation counts}

\item{decomposed_signature_matrix}{Posterior signature probability per nucleotide (Summing across channels should add up to 1)}

\item{signature_reference_matrix}{Signature channel emission probabilities (Summing across signature should add up to 1)}
}
\value{
Updated metadata with coverage metrics added.
}
\description{
Add Cosine Similarity to Metadata
}
\details{
Calculates and adds cosine similarity metrics for mutational signatures to the metadata.
Cosine similarity is computed between the decomposed signature matrix and the reference signature matrix.
}
