% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_signatures}
\alias{add_signatures}
\title{Add Signatures}
\usage{
add_signatures(
  metadata,
  activities_sbs_signatures,
  activities_indel_signatures,
  deconstructsigs_sbs_signatures
)
}
\arguments{
\item{metadata}{A data.table containing metadata.}

\item{activities_sbs_signatures}{Activities of SBS signatures.}

\item{activities_indel_signatures}{Activities of indel signatures.}

\item{deconstructsigs_sbs_signatures}{DeconstructSigs SBS signatures.}

\item{signatures_pair_name}{The name of the signature pair.}
}
\value{
Updated metadata with mutational signatures added.
}
\description{
Adds mutational signatures to the metadata based on various signature matrices.
}
