% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_sv_counts}
\alias{add_sv_counts}
\title{Add Structural Variant Counts}
\usage{
add_sv_counts(metadata, jabba_gg = NULL)
}
\arguments{
\item{jabba_gg}{Path to JaBbA graph RDS file}

\item{meta_dt}{A data.table containing metadata}
}
\value{
Updated metadata with SV counts
}
\description{
Adds junction, loose end, and total SV counts from JaBbA graph
}
