% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_sv_types}
\alias{add_sv_types}
\title{Add Structural Variant Types}
\usage{
add_sv_types(metadata, jabba_gg = NULL, complex = NULL)
}
\arguments{
\item{metadata}{A data.table containing metadata}

\item{jabba_gg}{Path to JaBbA graph RDS file}

\item{complex}{Path to complex events RDS file}
}
\value{
Updated metadata with SV type counts
}
\description{
Adds counts of different SV types from JaBbA graph and complex events
}
