% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_variant_counts}
\alias{add_variant_counts}
\title{Add Variant Counts}
\usage{
add_variant_counts(metadata, somatic_snvs = NULL, genome = "hg19")
}
\arguments{
\item{metadata}{A data.table containing metadata.}

\item{somatic_snvs}{Path to somatic SNV VCF file}

\item{genome}{The genome reference used.}
}
\value{
Updated metadata with variant counts added.
}
\description{
Adds variant counts to the metadata based on somatic and germline SNVs.
}
