% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_events}
\alias{aggregate_events}
\title{aggregate_events}
\usage{
aggregate_events(pair, events_tbl, jabba_gg, cohort_type = "paired")
}
\arguments{
\item{pair}{patient id}
}
\value{
data.table or NULL
}
\description{
Aggregate events per gene. 
Downstream of create_filtered_events().
}
