% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotype.R
\name{annotate_karyotype}
\alias{annotate_karyotype}
\title{Annotate Karyotype}
\usage{
annotate_karyotype(
  gg,
  segment_size_threshold = 5e+06,
  chr_fraction_threshold = 0.9,
  arm_fraction_threshold = 0.9,
  band_segment_ratio_threshold = 0.1
)
}
\description{
Convert gGraph to karyotype
}
