% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{arrow_temp}
\alias{arrow_temp}
\title{arrow_temp}
\usage{
arrow_temp(
  patient_id = NA,
  order = NA,
  x = list(NA),
  ref = NA,
  chart_type = "scatterplot",
  visible = TRUE,
  title = NA,
  type = "scatterplot",
  field = "foreground",
  overwrite = FALSE
)
}
\arguments{
\item{patient_id}{patient.id to be added to pgvdb}

\item{order}{optional entry if you order plots with a column order}

\item{x}{path to a granges rds or a granges object}

\item{ref}{reference to use for pgvdb}

\item{chart_type}{defaultChartType for pgvdb, default is scatterplot, can also be area}

\item{visible}{TRUE/FALSE whether the plot is hidden or showing in pgv}

\item{title}{title of the plot in pgvdb}

\item{type}{scatterplot, should not change for this plot type}

\item{field}{field to plot as the y value, default is foreground}

\item{overwrite}{TRUE/FALSE to overwrite an existing arrow}
}
\description{
function to create data.table for arrow files for pgvdb
}
\author{
Stanley Clarke
}
