% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{assign_in_namespace}
\alias{assign_in_namespace}
\title{Version of utils::assignInNamespace}
\usage{
assign_in_namespace(x, value, ns, pos = -1, envir = as.environment(pos))
}
\description{
can be used to reassign function into a namespace
USE WITH CAUTION
}
