% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bw_temp}
\alias{bw_temp}
\title{bw_temp}
\usage{
bw_temp(
  patient_id = NA,
  order = NA,
  x = list(NA),
  ref = NA,
  chart_type = "area",
  visible = TRUE,
  title = NA,
  type = "bigwig",
  field = "foreground",
  overwrite = FALSE
)
}
\arguments{
\item{patient_id}{patient.id to be added to pgvdb}

\item{order}{optional entry if you order plots with a column order}

\item{x}{path to a granges rds or a granges object}

\item{ref}{reference to use for pgvdb}

\item{chart_type}{defaultChartType for pgvdb, default is area, can also be scatterplot}

\item{visible}{TRUE/FALSE whether the plot is hidden or showing in pgv}

\item{title}{title of the plot in pgvdb}

\item{type}{bigwig, should not change for this plot type}

\item{field}{field to plot as the y value, default is foreground}

\item{overwrite}{TRUE/FALSE to overwrite an existing bw}
}
\description{
function to create data.table for bigwig files for pgvdb
}
\author{
Stanley Clarke
}
