% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseReportsData.R
\name{CaseReportsData}
\alias{CaseReportsData}
\title{CaseReportsData Class}
\description{
A class to handle case reports data
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{case_reports}}{data.table containing case reports data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CaseReportsData-new}{\code{CaseReportsData$new()}}
\item \href{#method-CaseReportsData-clone}{\code{CaseReportsData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CaseReportsData-new"></a>}}
\if{latex}{\out{\hypertarget{method-CaseReportsData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CaseReportsData object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaseReportsData$new(case_reports_data_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Character string specifying path to case reports data directory}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new CaseReportsData object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CaseReportsData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CaseReportsData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CaseReportsData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
