% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlights.R
\name{change_names}
\alias{change_names}
\title{Change names of data.frame/matrix}
\usage{
change_names(obj, old, new)
}
\description{
Mean to be an easy wrapper that uses either two
equal length character vectors, or a single named
character vector to rename the columns of a data frame
or matrix.
}
