% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-number-graph.R
\name{collapse_allelic_ggraph}
\alias{collapse_allelic_ggraph}
\title{Collapse Allelic graph}
\usage{
collapse_allelic_ggraph(gg)
}
\description{
Reduce number of segments by ignoring phase switches.
And redundant segments -- adjacent segments with the same
major and minor allele copy numbers
}
