% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncotable.R
\name{collect_copy_number_jabba}
\alias{collect_copy_number_jabba}
\title{collect_copy_number_jabba}
\usage{
collect_copy_number_jabba(
  jabba_rds,
  pge,
  amp.thresh,
  del.thresh,
  verbose = TRUE,
  karyograph = NULL
)
}
\arguments{
\item{jabba_rds}{Path to the jabba.simple.rds file.}

\item{pge}{GRanges object with gencode annotations.}

\item{amp.thresh}{SCNA amplification threshold.}

\item{del.thresh}{SCNA deletion threshold.}

\item{verbose}{Logical flag to indicate if messages should be printed.}

\item{karyograph}{Optional path to the karyograph.rds file.}
}
\value{
A data.table containing processed copy number and jabba information.
}
\description{
Collects copy number and jabba data from a specified file and processes it.
}
