% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{collect_gene_fusions}
\alias{collect_gene_fusions}
\title{collect_gene_fusions}
\usage{
collect_gene_fusions(fusions, pge, verbose = TRUE)
}
\arguments{
\item{fusions}{Path to the fusions.rds file.}

\item{pge}{GRanges object with gencode annotations.}

\item{verbose}{Logical flag to indicate if messages should be printed.}
}
\value{
A data.table containing processed gene fusion information.
}
\description{
Collects gene fusion data from a specified file and processes it.
}
