% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{collect_gene_mutations}
\alias{collect_gene_mutations}
\title{collect_gene_mutations}
\usage{
collect_gene_mutations(annotated_bcf, jabba_rds, filter, verbose = TRUE)
}
\arguments{
\item{annotated_bcf}{Path to the annotated.bcf file.}

\item{jabba_rds}{Path to the jabba.simple.rds file.}

\item{filter}{Filter to apply to the variants.}

\item{verbose}{Logical flag to indicate if messages should be printed.}
}
\value{
A data.table containing processed gene mutation information.
}
\description{
Collects gene mutation data from a specified file and processes it.
}
