% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{collect_oncokb}
\alias{collect_oncokb}
\title{collect_oncokb}
\usage{
collect_oncokb(oncokb_maf, multiplicity = NA_character_, verbose = TRUE)
}
\arguments{
\item{oncokb_maf}{Path to the oncokb MAF file.}

\item{multiplicity}{Path to the multiplicity file.}

\item{verbose}{Logical flag to indicate if messages should be printed.}
}
\value{
A data.table containing processed OncoKB mutation information.
}
\description{
Collects OncoKB mutation data from a specified file and processes it.
}
