% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{collect_oncokb_fusions}
\alias{collect_oncokb_fusions}
\title{collect_oncokb_fusions}
\usage{
collect_oncokb_fusions(oncokb_fusions, pge, cytoband, verbose = TRUE)
}
\arguments{
\item{oncokb_fusions}{Path to the oncokb fusions file.}

\item{verbose}{Logical flag to indicate if messages should be printed.}
}
\value{
A data.table containing processed OncoKB Fusion information.
}
\description{
Collects OncoKB Fusion data from a specified file and processes it.
}
\author{
Kevin Hadi
}
