% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{color2numeric}
\alias{color2numeric}
\title{Converts a color from hexadecimal to numeric format}
\usage{
color2numeric(color)
}
\arguments{
\item{color}{A character vector of colors in hexadecimal format (e.g., "#FF0000").}
}
\value{
A numeric vector representing the colors.
}
\description{
Converts a color from hexadecimal to numeric format
}
