% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{compute_signature_averages}
\alias{compute_signature_averages}
\title{Compute Signature Averages}
\usage{
compute_signature_averages(
  sig_file,
  is_indel = FALSE,
  is_deconstruct_sigs = FALSE
)
}
\arguments{
\item{sig_file}{Path to signature file}

\item{is_indel}{Boolean indicating if processing indel signatures}

\item{is_deconstruct_sigs}{Boolean indicating if using deconstructSigs format}

\item{sample}{Sample or pair name needed to subset to the correct sample}
}
\value{
List of signature averages
}
\description{
Computes signature averages from various signature files
}
