% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\docType{data}
\name{config_parameter_names}
\alias{config_parameter_names}
\title{Configuration parameter names}
\format{
An object of class \code{character} of length 15.
}
\usage{
config_parameter_names
}
\description{
List of configuration parameter column names
}
\details{
This variable contains the names of columns that are used
for configuration parameters in the Cohort inputs data.table
}
\keyword{datasets}
