% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{convert_json_to_arrow}
\alias{convert_json_to_arrow}
\title{Convert a JSON file to an Arrow file.}
\usage{
convert_json_to_arrow(json_file_path, arrow_file_path = NULL)
}
\arguments{
\item{json_file_path}{Path to the input JSON file.}

\item{arrow_file_path}{Optional. Path to the output Arrow file.
If NULL, the output path will be the same as the input JSON file,
but with the .arrow extension.}
}
\value{
Invisibly returns the path to the created Arrow file.
}
\description{
Convert a JSON file to an Arrow file.
}
\author{
Shihab Dider
}
