% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cov2arrow_pgv}
\alias{cov2arrow_pgv}
\title{cov2arrow_pgv}
\usage{
cov2arrow_pgv(
  patient.id,
  dryclean_cov,
  jabba_gg = NULL,
  purity = NULL,
  ploidy = NULL,
  mask = NULL,
  ref,
  title = NA,
  seq.fix = NULL,
  chart_type = "scatterplot",
  visible = TRUE,
  field = "foreground",
  new_col = "foregroundabs",
  overwrite = FALSE,
  order = NA,
  binsize = 10000
)
}
\arguments{
\item{patient.id}{patient name to add to pgv}

\item{dryclean_cov}{dryclean coverage}

\item{jabba_gg}{optional jabba ggraph. If null, needs purity or ploidy}

\item{purity}{optional purity value. If null needs ggraph}

\item{ploidy}{optional ploidy value. If null needs ggraph}

\item{mask}{optional mask to use after rebinning}

\item{ref}{reference to use for pgvdb}

\item{title}{optional title for plot for pgvdb}

\item{seq.fix}{optional seqlengths vector to fix granges seqlengths}

\item{chart_type}{default chart type for plot in pgvdb. scatterplot or area}

\item{visible}{whether the plot is visible in pgv}

\item{field}{column to do rel2abs on}

\item{new_col}{new column after rel2abs}

\item{overwrite}{whether to overwrite the current bigwig}

\item{order}{optional order if using a column order in your pgvdb object to sort}

\item{binsize}{size to rebin coverages, default 1e4}
}
\value{
NULL or segstats table
}
\description{
function to create arrow with rel2abs from coverage file. Can take a ggraph or specified purity and ploidy to calculate rel2abs
}
\author{
Stanley Clarke
}
