% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cov_abs}
\alias{cov_abs}
\alias{cov2abs}
\title{cov_abs}
\usage{
cov2abs(
  dryclean_cov,
  jabba_gg = NULL,
  purity = NULL,
  ploidy = NULL,
  field = "foreground",
  new_col = "foregroundabs"
)
}
\arguments{
\item{dryclean_cov}{dryclean coverage}

\item{jabba_gg}{optional jabba ggraph. If null, needs purity or ploidy}

\item{purity}{optional purity value. If null needs ggraph}

\item{ploidy}{optional ploidy value. If null needs ggraph}

\item{field}{column in granges to convert with rel2abs}

\item{new_col}{new column to add to add for the converted rel2abs}
}
\value{
NULL or segstats table
}
\description{
function to run rel2abs on coverage taking either a ggraph or purity & ploidy
}
\author{
Stanley Clarke
}
