% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-width-distribution.R
\name{create_2d_purity_ploidy}
\alias{create_2d_purity_ploidy}
\title{2D QC Purity Ploidy Plot}
\usage{
create_2d_purity_ploidy(
  cohorttuple,
  mask = TRUE,
  mask_gr = system.file("extdata", "data", "maskA_re.rds", package = "Skilift"),
  subsample_per_group = 1000,
  datadir
)
}
\description{
Create scatter plot of copy number fits
}
\details{
Data comes from multiplicity
}
\author{
Kevin Hadi

Aditya Deshpande
}
