% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_distributions}
\alias{create_distributions}
\title{create_distributions}
\usage{
create_distributions(
  case_reports_datadir,
  filter_patients = NULL,
  cores = 1,
  write_to_json = FALSE,
  common_dir = NULL,
  overwrite = FALSE,
  haveSignatures = TRUE
)
}
\arguments{
\item{case_reports_datadir}{directory containing all patient directories}

\item{filter_patients}{list() - subset of samples to filter on to create the distributions}

\item{cores}{number of cores to use for parallel processing}

\item{write_to_json}{boolean flag to indicate whether to write the files to JSON or not}

\item{common_dir}{path to the common directory in which to write the distribution JSONs (required if write_to_json is TRUE)}

\item{overwrite}{boolean flag to indicate whether to overwrite existing files (required if write_to_json is TRUE)}
}
\value{
List of data tables containing distributions if write_to_json is FALSE, otherwise NULL
}
\description{
Creates the background distributions for aggregated metadata KPIs across all patients
}
