% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{create_filtered_events}
\alias{create_filtered_events}
\title{create_filtered_events}
\usage{
create_filtered_events(
  pair,
  oncotable,
  jabba_gg,
  out_file,
  temp_fix = FALSE,
  return_table = FALSE,
  cohort_type = "paired"
)
}
\arguments{
\item{pair}{patient id}

\item{oncotable}{oncotable task output}

\item{jabba_gg}{JaBbA output ggraph or complex}

\item{out_file}{path to write json}

\item{temp_fix}{TRUE/FALSE whether to apply temporary fix}

\item{return_table}{TRUE/FALSE whether to return the data.table}
}
\value{
data.table or NULL
}
\description{
Create filtered events for a single sample
}
