% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlights.R
\name{create_heme_highlights}
\alias{create_heme_highlights}
\title{create_heme_highlights}
\usage{
create_heme_highlights(
  events_tbl,
  jabba_gg,
  out_file,
  hemedb_path =
    "/gpfs/data/imielinskilab/projects/Clinical_NYU/db/master_heme_database.20250128_095937.790322.rds",
  duncavage_path = "~/projects/Clinical_NYU/db/Duncavage_Blood_22.rds"
)
}
\description{
Create highlighted heme events for a single sample. 
This code sits downstream of create_filtered_events
and is used within the lift_filtered_events method.
}
