% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlights.R
\name{create_heme_highlights_eln_risk_score}
\alias{create_heme_highlights_eln_risk_score}
\title{ELN Risk Score}
\usage{
create_heme_highlights_eln_risk_score(
  small_muts,
  svs,
  cna,
  karyotype_list,
  jabba_gg
)
}
\description{
Logic adapted from Fig 2 of Lachowiez et al. Blood Adv 2022
}
\details{
Heuristic for risk score based on small mutations, 
fusion genes (not taken from karyotypes, but the gene level annotations), 
and arm level CNA (taken from karyotypes).
}
