% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{create_mutations_catalog}
\alias{create_mutations_catalog}
\title{create_mutations_catalog}
\usage{
create_mutations_catalog(sigmat_path, is_indel = FALSE)
}
\arguments{
\item{sigmat_path}{Path to either the SBS or ID matrix}

\item{is_indel}{Boolean to indicate if the matrix is for indels}
}
\value{
list containing:
  - data: data.frame with columns:
    - id: row identifier
    - insdel/tnc: mutation type (insdel for indels, tnc for SBS)
    - mutations: count of mutations
}
\description{
Creates a mutations catalog data frame from sigprofiler matrix generator output
}
\author{
Shihab Dider, Sukanya Panja
}
