% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_mutations_catalog_json}
\alias{create_mutations_catalog_json}
\title{create_mutations_catalog_json}
\usage{
create_mutations_catalog_json(sig_matrix_path, is_indel, output_dir)
}
\arguments{
\item{sig_matrix_path}{Path to either the SBS or ID matrix}

\item{is_indel}{Boolean to indicate if the matrix is for indels}

\item{output_dir}{directory to write the jsons, will create directory for each patient id}
}
\description{
Builds the mutations catalog for case reports using the output of sigprofiler matrix generator
}
\author{
Shihab Dider, Sukanya Panja
}
