% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{create_oncotable}
\alias{create_oncotable}
\title{create_oncotable}
\usage{
create_oncotable(
  cohort,
  amp_thresh_multiplier = 1.5,
  gencode = "/gpfs/data/imielinskilab/DB/GENCODE/gencode.v19.annotation.gtf.nochr.rds",
  cytoband = system.file("extdata", "data", "cytoband.rds", package = "Skilift"),
  outdir,
  cores = 1
)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{amp_thresh_multiplier}{amp.thresh for oncotable is amp_thresh_multiplier*ploidy}

\item{gencode}{file to gencode annotations (uses v29lift37 by default)}

\item{outdir}{path to directory in which to write oncotable outputs}

\item{cores}{number of cores for parallel processing}
}
\value{
None
}
\description{
function to create oncotable for use with filtered_events_json
}
\author{
Shihab Dider, Joel Rosiene
}
