% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-width-distribution.R
\name{create_pp_plot}
\alias{create_pp_plot}
\title{Purity Ploidy Plot}
\usage{
create_pp_plot(
  jabba_gg = NA,
  het_pileups = NULL,
  seg.fname = NA,
  field = "count",
  purity = NA,
  ploidy = NA,
  plot.min = -2,
  plot.max = 2,
  bins = 500,
  height = 800,
  width = 800,
  output.fname = "./",
  is.wgs = FALSE
)
}
\description{
Lorem Ipsem
}
\author{
Aditya Deshpande
}
