% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlights.R
\name{create_summary}
\alias{create_summary}
\title{create_summary}
\usage{
create_summary(
  events_tbl,
  altered_copies_threshold = 0.1,
  cohort_type,
  cohorttuple
)
}
\description{
Create high level summary of
mutations that will get passed
to metadata.json.
}
