% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant-qc.R
\name{create_variant_qc}
\alias{create_variant_qc}
\title{create_variant_qc}
\usage{
create_variant_qc(somatic_snvs, reference_name = "hg19")
}
\arguments{
\item{somatic_snvs}{Path to sage VCF file to be used (paired T-N/ Tumor only)}

\item{reference_name}{Reference genome name (e.g., "hg19", "hg38")}
}
\value{
data.table containing variant QC metrics
}
\description{
Function to create quality control metrics for somatic variants
}
\author{
Shihab Dider, Tanubrata Dey, Stanley Clarke
}
