% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{dcast_skilift}
\alias{dcast_skilift}
\title{Cast table}
\usage{
dcast_skilift(
  tbl,
  id_columns,
  type_columns,
  cast_columns,
  keep_remaining = FALSE,
  sep_cast = "___",
  prefix_type = TRUE,
  drop = FALSE,
  use_regex = FALSE
)
}
\description{
Using base R for robustness
}
