% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dcastski}
\alias{dcastski}
\title{Cast table}
\usage{
dcastski(
  tbl,
  id_columns,
  type_columns,
  cast_columns,
  keep_remaining = FALSE,
  sep_cast = "___",
  prefix_type = TRUE,
  drop = FALSE,
  use_regex = FALSE
)
}
\description{
Cast using base R.
}
\details{
Using base R for robustness and flexibility.
}
