% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{DEFAULT_JABBA}
\alias{DEFAULT_JABBA}
\title{Default Jabba Type}
\usage{
DEFAULT_JABBA(object, get_object = "cohort", test_mode = all, verbose = TRUE)
}
\description{
Get correct jabba column to use globally across all methods
}
