% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dt2json_mut}
\alias{dt2json_mut}
\title{dt2json_mut}
\usage{
dt2json_mut(
  dt,
  patient.id,
  ref,
  settings,
  file_name = paste(getwd(), "test.json", sep = "/"),
  meta_data = NULL,
  y_col = NULL
)
}
\arguments{
\item{dt}{data.table with seqnames,start and end}

\item{patient.id}{patient id to be added to pgvdb or case reports}

\item{ref}{reference for pgv or case reports}

\item{file_name}{the file the json should be written to}

\item{meta_data}{extra information to add to the json}

\item{y_col}{column to be used for y axis values}
}
\description{
function to create a mutation json, used in create_somatic_json
}
\author{
Stanley Clarke
}
