% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skilift.R
\docType{data}
\name{empty_plot_default}
\alias{empty_plot_default}
\title{Global Variables}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 0 rows and 11 columns.
}
\usage{
empty_plot_default
}
\description{
Global Variables
}
\keyword{datasets}
