% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Extract Metrics from QC File}
\usage{
extract_metrics(qc_data, metrics, pair)
}
\arguments{
\item{qc_data}{Data.table containing QC metrics}

\item{metrics}{Named vector of column names to extract}

\item{pair}{Sample pair identifier}
}
\value{
Data.table with extracted metrics
}
\description{
Helper function to extract specified metrics from a QC file
}
