% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\docType{data}
\name{GENCODE_DEFAULTS}
\alias{GENCODE_DEFAULTS}
\title{Default gencode paths}
\format{
An object of class \code{list} of length 2.
}
\usage{
GENCODE_DEFAULTS
}
\description{
Provided as an exported variable
}
\details{
Providing gencode defaults as a variable exposed to user
}
\keyword{datasets}
