% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{get_default_gencode}
\alias{get_default_gencode}
\title{Get default gencode}
\usage{
get_default_gencode(
  gencode_path_defaults = Skilift::GENCODE_DEFAULTS,
  assembly = NULL
)
}
\description{
Function to grab default gencode
}
\details{
Use environment variables
GENCODE_PATH
DEFAULT_ASSEMBLY
To get appropriate gencode and
assign it to a package level variable.
}
