% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{get_gene_ampdels_from_jabba}
\alias{get_gene_ampdels_from_jabba}
\title{Get gene amplifications and deletions from jabba object}
\usage{
get_gene_ampdels_from_jabba(
  jab,
  pge,
  amp.thresh = 4,
  del.thresh = 0.5,
  nseg = NULL,
  min_cn_quantile_threshold = 0.1,
  max_cn_quantile_threshold = 0.9
)
}
\arguments{
\item{jab}{character path to jabba file or gGraph object}
}
\description{
Copied from github::mskilab-org/skitools.
This function takes in a jabba object and returns
amplifications and deletions
}
