% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-width-distribution.R
\name{get_segstats}
\alias{get_segstats}
\title{create_ppfit_json}
\usage{
get_segstats(
  balanced_jabba_gg,
  tumor_coverage,
  coverage_field = "foreground",
  settings = Skilift:::default_settings_path,
  ref = "hg19",
  max_na = 0.9,
  cores = 1
)
}
\arguments{
\item{balanced_jabba_gg}{the balanced_gg ggraph or the path to it}

\item{tumor_coverage}{path to associated coverage, if null, will try to pull path from pairs table}

\item{coverage_field}{field in coverage file to use for segstats}

\item{settings}{path to settings.json (from PGV)}

\item{ref}{reference name}

\item{max_na}{max.na used with JaBbA}

\item{cores}{cores for JaBbA:::segstats}
}
\value{
segstats table
}
\description{
function to create segmentation plots in case reports
}
\author{
Stanley Clarke, Tanubrata Dey, Shihab Dider
drcln_cov = readRDS(thisp$decomposed_cov[i])
}
