% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyotype.R
\name{grl2bedpe}
\alias{grl2bedpe}
\title{SV in GRangesList to bedpe}
\usage{
grl2bedpe(
  grl,
  add_breakend_mcol = FALSE,
  flip = FALSE,
  as.data.table = TRUE,
  zerobased = TRUE
)
}
\description{
Convert GRangesList representation of SVs to bedpe-like format
}
