% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift-wrappers.R
\name{has_required_columns}
\alias{has_required_columns}
\title{lift wrapper helper function}
\usage{
has_required_columns(cohort, columns, any = FALSE)
}
\description{
Helper function to check if required columns exist
}
