% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{indels_sigprofiler2meta}
\alias{indels_sigprofiler2meta}
\title{indels_sigprofiler2meta}
\usage{
indels_sigprofiler2meta(indel_file, sample)
}
\arguments{
\item{indel_file}{path to Assignment_Solution_Activities.txt, output of sigprofiler}

\item{sample}{sample or pair name-needed to subset to the correct sample}
}
\description{
function to generate metadata indel signature list from sigprofiler output
}
\author{
Sukanya Panja, Stanley Clarke
}
