% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{initialize_metadata_columns}
\alias{initialize_metadata_columns}
\title{Initialize Metadata Columns}
\usage{
initialize_metadata_columns(pair)
}
\arguments{
\item{pair}{The sample pair identifier}
}
\value{
A data.table with all possible columns initialized to NA
}
\description{
Creates a data.table with all expected columns initialized to NA
}
