% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{is_loosely_na}
\alias{is_loosely_na}
\title{Kind of NA}
\usage{
is_loosely_na(
  values,
  character_nas = c("NA", "na", "N/A", "n/a", "NULL", "Null", "null"),
  other_nas = NULL
)
}
\description{
Test for NAs of different types
}
\details{
NA may not be the only type that we want to test for

If it's a character, we may want to loosely test for other "NAs"
}
