% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-width-distribution.R
\name{lift_2d_purity_ploidy_plot}
\alias{lift_2d_purity_ploidy_plot}
\title{Lift 2d purity ploidy}
\usage{
lift_2d_purity_ploidy_plot(cohort, output_data_dir, cores = 1)
}
\description{
Lift 2d purity ploidy
}
\author{
Kevin Hadi

Aditya Deshpande
}
