% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift-wrappers.R
\name{lift_all}
\alias{lift_all}
\title{Run all lift methods on a cohort}
\usage{
lift_all(
  cohort,
  output_data_dir,
  cores = 1,
  settings = Skilift:::internal_settings_path,
  max_cn = 100,
  annotations = NULL,
  coverage_field = "foreground",
  color_field = NULL,
  bin.width = 10000,
  node_metadata = c("gene", "feature_type", "annotation", "REF", "ALT", "variant.c",
    "variant.p", "vaf", "transcript_type", "impact", "rank"),
  multiplicity_field = "altered_copies",
  genome_length = c(1:22, "X", "Y"),
  path_to_nodejs = "/gpfs/share/apps/nodejs/22.9.0/bin/node",
  ...
)
lift_all(cohort, output_data_dir, cores = 1, ...)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}

\item{...}{Additional arguments}
}
\value{
None
}
\description{
This function runs all available lift methods on a cohort to generate the complete
set of output files needed for visualization.
}
