% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-width-distribution.R
\name{lift_allelic_pp_fit}
\alias{lift_allelic_pp_fit}
\title{lift_allelic_pp_fit}
\usage{
lift_allelic_pp_fit(
  cohort,
  output_data_dir,
  cores = 1,
  file.name = "hetsnps_major_minor.png"
)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}

\item{file.name}{Name of the output file (default: "hetsnps_major_minor.png")}
}
\description{
This function will create PNG files for allelic ploidy/purity fits
}
\author{
Johnathan Rafailov
}
