% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam.R
\name{lift_bam}
\alias{lift_bam}
\title{lift_bam}
\usage{
lift_bam(cohort, output_data_dir, cores = 1, overwrite = FALSE)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}
}
\value{
None
}
\description{
Lift bam file to output data directory.
Currently hardlinking.
}
