% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-number-graph.R
\name{lift_copy_number_graph}
\alias{lift_copy_number_graph}
\title{lift_copy_number_graph}
\usage{
lift_copy_number_graph(
  cohort,
  settings = internal_settings_path,
  output_data_dir,
  is_allelic = FALSE,
  max_cn = 100,
  annotations = list(c("bfb", "chromoplexy", "chromothripsis", "del", "dm", "cpxdm",
    "dup", "pyrgo", "rigma", "simple", "tic", "tyfonas")),
  cores = 1
)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{is_allelic}{Boolean indicating whether to process allelic (TRUE) or total (FALSE) copy number}

\item{cores}{Number of cores for parallel processing (default: 1)}
}
\value{
None
}
\description{
Create copy number graph JSON files for all samples in a cohort
}
