% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-width-distribution.R
\name{lift_coverage_jabba_cn}
\alias{lift_coverage_jabba_cn}
\title{lift_coverage_jabba_cn}
\usage{
lift_coverage_jabba_cn(
  cohort,
  output_data_dir,
  cores = 1,
  width = 10000,
  mask_path = system.file("extdata", "data", "maskA_re.rds", package = "Skilift"),
  mask = TRUE,
  save_html = TRUE,
  save_png = TRUE,
  save_data = TRUE
)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}

\item{width}{tile width of the coverage GRanges  (default: 10kbp)}

\item{mask}{logical value to mask the data or not; default is TRUE}

\item{save_html}{Logical value to save HTML files (default: TRUE)}

\item{save_png}{Logical value to save PNG files (default: TRUE)}

\item{save_data}{Logical value to save data files (default: TRUE)}
}
\value{
None
}
\description{
This function will create a boxplot and calculate the linear correlation between the foreground from coverage GRanges and JaBbA graph CN.
}
\author{
Johnathan Rafailov
}
