% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{lift_coverage_track}
\alias{lift_coverage_track}
\title{lift_coverage_track}
\usage{
lift_coverage_track(
  cohort,
  output_data_dir,
  cores = 1,
  cohort_column = "tumor_coverage",
  coverage_field = "foreground",
  color_field = NULL,
  bin.width = NA_integer_
)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}

\item{cohort_column}{Field in cohort inputs table that specifies the source data}

\item{coverage_field}{Field in the input GRanges object that specifies the values to write to arrow}

\item{color_field}{Field in the input GRanges object that specifies the color values. Note this must be a hexadecimal value.}
}
\value{
None
}
\description{
Create coverage arrow files for all samples in a cohort.
Inputs can be any GRanges with a color field
}
