% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{lift_datafiles_json}
\alias{lift_datafiles_json}
\title{lift_datafiles_json}
\usage{
lift_datafiles_json(output_data_dir)
}
\arguments{
\item{data_dir}{Directory containing data files}
}
\value{
None
}
\description{
Create a combined JSON file for all data files in a directory
}
