% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{lift_denoised_coverage}
\alias{lift_denoised_coverage}
\title{lift_denoised_coverage}
\usage{
lift_denoised_coverage(cohort, output_data_dir, cores = 1, mask = TRUE)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}

\item{mask}{Logical indicating whether to mask the coverage data (default: TRUE)}
}
\value{
None
}
\description{
Create denoised coverage arrow files for all samples in a cohort
}
