% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{lift_filtered_events}
\alias{lift_filtered_events}
\title{lift_filtered_events}
\usage{
lift_filtered_events(cohort, output_data_dir, cores = 1, in_heme_mode = FALSE)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}
}
\value{
None
}
\description{
Create filtered events for all samples in a cohort
}
