% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{lift_hetsnps}
\alias{lift_hetsnps}
\title{lift_hetsnps}
\usage{
lift_hetsnps(cohort, output_data_dir, cores = 1)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}
}
\value{
None
}
\description{
Create hetsnps arrow files for all samples in a cohort
}
