% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{lift_metadata}
\alias{lift_metadata}
\title{lift_metadata}
\usage{
lift_metadata(cohort, output_data_dir, cores = 1, genome_length = NULL)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}

\item{genome_length}{Genome length for the samples (for targeted panels or WES data)}
}
\value{
None
}
\description{
Create metadata files for all samples in a cohort
}
