% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicity.R
\name{lift_multiplicity}
\alias{lift_multiplicity}
\title{lift_multiplicity}
\usage{
lift_multiplicity(
  cohort,
  is_germline = FALSE,
  node_metadata = c("Variant_Classification", "Gene", "HGVSc", "HGVSp", "variant.g",
    "vaf", "alt", "ref", "normal.alt", "normal.ref", "FILTER", "ONCOGENIC",
    "MUTATION_EFFECT", "HIGHEST_LEVEL"),
  output_data_dir,
  field = "altered_copies",
  cores = 1
)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{is_germline}{Logical indicating if mutations are germline (default: FALSE)}

\item{node_metadata}{Additional columns to include in node data (default: c("gene", "feature_type", "annotation", "REF", "ALT", "variant.c", "variant.p", "vaf", "transcript_type", "impact", "rank"))}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}
}
\value{
None
}
\description{
Create multiplicity JSON files for all samples in a cohort
}
