% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-width-distribution.R
\name{lift_multiplicity_fits}
\alias{lift_multiplicity_fits}
\title{lift_multiplicity_fits}
\usage{
lift_multiplicity_fits(cohort, output_data_dir, cores = 1)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}
}
\description{
This function will create JSON files for somatic, germline, and hetsnps multiplicity fits
-somatic: altered_copies and total_snv_copies
-germline: altered_copies and total_snv_copies
-hetsnps: major_snv_copies and minor_snv_copies
}
