% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-width-distribution.R
\name{lift_pp_plot}
\alias{lift_pp_plot}
\title{Lift 2d purity ploidy plot}
\usage{
lift_pp_plot(cohort, output_data_dir, cores = 1)
}
\description{
Lift 2d purity ploidy plot
}
\author{
Aditya Deshpande
}
