% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-width-distribution.R
\name{lift_purple_sunrise_plot}
\alias{lift_purple_sunrise_plot}
\title{lift_purple_sunrise_plot}
\usage{
lift_purple_sunrise_plot(
  cohort,
  output_data_dir,
  cores = 1,
  save_pngs = TRUE,
  save_html = TRUE,
  save_data = TRUE
)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}

\item{save_pngs}{Logical value to save PNG files (default: TRUE)}

\item{save_html}{Logical value to save HTML files (default: TRUE)}

\item{save_data}{Logical value to save data files (default: TRUE)}
}
\description{
This function will create JSON files for purple purity/ploidy sunrise plots
}
\references{
Code adapted from:
 - https://github.com/hartwigmedical/hmftools/tree/642436265858083a0bfc81b793a51ccde42edd02/purple/src/main/resources/r/copyNumberPlots.R
}
\author{
Johnathan Rafailov
}
