% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-cell.R
\name{lift_sc_genome_graphs}
\alias{lift_sc_genome_graphs}
\title{lift_sc_genome_graphs}
\usage{
lift_sc_genome_graphs(
  pgv,
  cells,
  gg_col,
  type = c("genome", "allelic"),
  ploidy_col = "ploidy",
  ref = "hg38",
  order = if (type == "genome") 10 else 15,
  visible = type == "genome",
  title = NULL,
  max.cn = 1000,
  link_allelic = TRUE,
  overwrite = FALSE,
  cores = 1
)
}
\arguments{
\item{pgv}{Skilift object}

\item{cells}{data.table with pair, the graph column and optionally the ploidy column}

\item{gg_col}{column of gGraph rds paths}

\item{type}{"genome" or "allelic"}

\item{ploidy_col}{column with cell ploidy, added to titles; NULL to omit}

\item{ref}{reference for new patients in PGV}

\item{order}{plot order}

\item{visible}{whether the plot is shown by default}

\item{title}{plot titles; default "<pair> Total CN|Allelic; Ploidy: x"}

\item{max.cn}{maximum copy number shown}

\item{link_allelic}{for type = "allelic", set allelicSource on each cell's genome plot}

\item{overwrite}{overwrite existing json files}

\item{cores}{number of cores}
}
\value{
the data.table of plots added (invisibly)
}
\description{
Add each cell's total copy-number (type = "genome") or allelic
(type = "allelic") JaBbA graph to a PGV instance with pgv$add_plots().
Allelic graphs are also linked to the cell's genome plot via allelicSource.
}
\author{
Stanley Clarke
}
