% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{lift_signatures}
\alias{lift_signatures}
\title{lift_signatures}
\usage{
lift_signatures(cohort, output_data_dir, cores = 1)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}
}
\value{
None
}
\description{
Create signature files for all samples in a cohort
}
