% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-width-distribution.R
\name{lift_somatic_fit}
\alias{lift_somatic_fit}
\title{lift_somatic_fit}
\usage{
lift_somatic_fit(
  cohort,
  output_data_dir,
  mask_path = system.file("extdata", "data", "maskA_re.rds", package = "Skilift"),
  mask = TRUE,
  bins = 1e+05,
  cores
)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{mask_path}{path to the mask file, contained in extdata of package itself; default is maskA as provided in the package}

\item{mask}{logical value to mask the data or not; default is TRUE}

\item{bins}{number of bins for histogram; should specify for lower limit to avoid performance issues; default = 100000}

\item{cores}{Number of cores for parallel processing (default: 1)}
}
\value{
None
}
\description{
a method to lift the somatic mutations as a histogram to JSON for viewing in gOS
}
\author{
Johnathan Rafailov
}
