% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{load_distributions}
\alias{load_distributions}
\title{load_distributions}
\usage{
load_distributions(common_dir, filter_patients = NULL)
}
\arguments{
\item{common_dir}{path to a folder to write all 7 jsons}

\item{filter_pateints}{list of samples to filter on to read in the distributions}
}
\description{
function to read in distributions files, same format returned as create_distributions with write_jsons = FALSE
}
\author{
Stanley Clarke
}
