% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{make_allelic_hets}
\alias{make_allelic_hets}
\alias{make_allelic_hetsnps}
\title{make_allelic_hets}
\usage{
make_allelic_hetsnps(
  het_pileups = NULL,
  min_normal_freq = 0.2,
  max_normal_freq = 0.8
)
}
\arguments{
\item{het_pileups}{(character) path to sites.txt with hetsnps}

\item{min_normal_freq}{(numeric) in [0, 1] min frequency in normal to count as het site}

\item{max_normal_freq}{(numeric) in [0, 1] max frequency in normal to count as het site}
}
\value{
GRanges with major and minor counts
}
\description{
returns melted hetsnps gRanges with allelic (major/minor) counts
}
\author{
Jonathan Rafailov, Shihab Dider
}
