% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{meltski}
\alias{meltski}
\title{Melt table}
\usage{
meltski(
  tbl,
  id.vars,
  measure.vars,
  is_measure_regex = FALSE,
  variable.name = "variable",
  value.name = "value",
  group_regex = "",
  group_select = "\\\\1",
  replacement_regex = "",
  replacement = "",
  drop = FALSE,
  keep_first_variable_col = FALSE,
  keep_remaining_cols = FALSE,
  keep_original_cols = FALSE,
  return_as_data_table = FALSE
)
}
\description{
Melt using base R.
}
\details{
Using base R for robustness and flexibility, not speed necessarily.
Since this will be used primarily for sample/pairs table manipulation.
}
