% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{merge.Cohort}
\alias{merge.Cohort}
\title{Merge Cohort objects}
\usage{
\method{merge}{Cohort}(x, y, ..., warn_duplicates = TRUE, rename_duplicates = FALSE)
}
\arguments{
\item{x}{Cohort object to merge}

\item{y}{Cohort object to merge}

\item{...}{Arbitrary remaining Cohort objects to merge}

\item{warn_duplicates}{Logical indicating whether to warn about duplicate pairs (default TRUE)}

\item{rename_duplicates}{Logical indicating whether to rename duplicate pairs instead of overwriting (default FALSE)}
}
\value{
A new Cohort object containing all data from input Cohorts
}
\description{
Combines two or more Cohort objects into a single Cohort
}
