% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge.repl}
\alias{merge.repl}
\title{merging data tables with collapsing columns with the same name}
\usage{
\method{merge}{repl}(
  dt.x,
  dt.y,
  sep = "_",
  replace_NA = TRUE,
  delete_xiny,
  force_y = TRUE,
  prefer_y,
  prefer_x,
  overwrite_x = FALSE,
  prefer_y_na,
  keep_order = FALSE,
  keep_colorder = TRUE,
  keep_factor = TRUE,
  suffixes = c(".x", ".y"),
  ...
)
}
\arguments{
\item{replace_NA}{logical, only use values in dt.y, any dt.x not in dt.y is clobbered (NA)}

\item{delete_xiny}{logical, delete columns in x that intersect with y columns (remove previous columns)}

\item{force_y}{logical, should x and y common columns be merged?}

\item{prefer_y}{logical, for x and y entries sharing common by keys, prefer y (logic to control whether to ignore na or not encoded by prefer_y_na)}

\item{overwrite_x}{logical, if force_y = TRUE, should NA values in y replace x?}

\item{prefer_y_na}{logical, for x and y entries sharing common by keys, prefer y and ignore NAs}
}
\value{
A data.table
}
\description{
Merge two data tables with various replacing strategies
for columns common between x and y that are not used to merge
(i.e. not specified in the "by" argument)
}
