% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{merge_oncokb_multiplicity}
\alias{merge_oncokb_multiplicity}
\title{Merge oncokb and multiplicity}
\usage{
merge_oncokb_multiplicity(oncokb, multiplicity, overwrite = FALSE)
}
\description{
Finds overlaps between oncokb maf coordinates and multiplicity
Oncokb is in MAF coordinates which is based on the altered bases,
while multiplicity encodes VCF coordinates based on reference bases.
}
