% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mutations_temp}
\alias{mutations_temp}
\title{mutations_temp}
\usage{
mutations_temp(
  patient_id = NA,
  order = NA,
  x = list(NA),
  ref = NA,
  source = "mutations.json",
  field,
  type = "mutations",
  visible = TRUE,
  title = NA,
  tag = NA,
  overwrite = FALSE
)
}
\arguments{
\item{patient_id}{patient.id to be added to pgvdb}

\item{order}{optional entry if you order plots with a column order}

\item{x}{mutations object as a list}

\item{ref}{reference to use for pgvdb}

\item{type}{walk, do not change for this plot type}

\item{visible}{TRUE/FALSE whether the plot is hidden or showing in pgv}

\item{title}{title of the plot in pgvdb}

\item{tag}{optional argument, can be binset to override y spacing in pgv}

\item{overwrite}{TRUE/FALSE to overwrite an existing genome json}

\item{annotation}{default is list of SVs, make null if no annotations present in object}
}
\description{
function to create data.table for mutations for pgvdb
}
\author{
Stanley Clarke
}
