% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{names2}
\alias{names2}
\title{robust name()}
\usage{
names2(x)
}
\arguments{
\item{str}{a path string}
}
\value{
a string with multiple parentheses replaced with a single parenthesis
}
\description{
gives back character vector same length of input regardless whether named or not
}
