% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\docType{data}
\name{nf_path_patterns}
\alias{nf_path_patterns}
\title{Default nf_modes for Cohort}
\format{
An object of class \code{list} of length 30.
}
\usage{
nf_path_patterns
}
\description{
Lorem ipsem
}
\details{
Lorem Ipsem
}
\keyword{datasets}
