% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skilift-package.R
\docType{data}
\name{PAIRED_READS_FACTOR}
\alias{PAIRED_READS_FACTOR}
\title{Paired reads factor}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
PAIRED_READS_FACTOR
}
\description{
Multiply paired reads by this number for coverage
}
\details{
To get binned coverage into base level coverage,
multiply by paired reads factor (2)
}
\keyword{datasets}
