% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{parse_oncokb_tier}
\alias{parse_oncokb_tier}
\title{Helper function to parse oncokb outputs
levels of evidence and assign tier. 
Tiering is simply:
1 = Clinically Actionable:
Therapeutic sensitivity, resistance,
diagnostic, or prognostic information is assigned
to the variant.
2 = Clinically Significant:
Clinically significant variant entails that the variant
does not have therapeutic, diagnostic, or prognostic effect
validated at FDA level, but is oncogenic.
3 = All others (VUS)}
\usage{
parse_oncokb_tier(
  oncokb,
  tx_cols = c("LEVEL_1", "LEVEL_2"),
  rx_cols = c("LEVEL_R1"),
  dx_cols = c("LEVEL_Dx1"),
  px_cols = c("LEVEL_Px1")
)
}
\arguments{
\item{oncokb}{data.table object holding oncokb outputs}
}
\description{
Helper function to parse oncokb outputs
levels of evidence and assign tier. 
Tiering is simply:
1 = Clinically Actionable:
Therapeutic sensitivity, resistance,
diagnostic, or prognostic information is assigned
to the variant.
2 = Clinically Significant:
Clinically significant variant entails that the variant
does not have therapeutic, diagnostic, or prognostic effect
validated at FDA level, but is oncogenic.
3 = All others (VUS)
}
\author{
Kevin Hadi
}
