% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{parse_pipeline_paths}
\alias{parse_pipeline_paths}
\title{nf-casereport Parser}
\usage{
parse_pipeline_paths(
  paths,
  initial_dt = data.table(pair = character(0)),
  path_patterns = Skilift::nf_path_patterns,
  id_name = "pair"
)
}
\description{
Parses outputs of the nf-casereport directory structure
}
\details{
This is a function that extracts paths from a 
pipeline. The default structure is that the
path patterns expected follow that of nf-casereport run.
The columns it extracts are defined by
variable Skilift::nf_column_map.
You can output the default template to modify it using
base::dput(Skilift::nf_path_patterns)
}
