% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ppfit_temp}
\alias{ppfit_temp}
\title{ppfit_temp}
\usage{
ppfit_temp(
  patient_id = NA,
  order = NA,
  x = list(NA),
  ref = NA,
  source = "ppfit.json",
  type = "ppfit",
  visible = TRUE,
  title = NA,
  annotation = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{patient_id}{patient.id to be added to pgvdb}

\item{order}{optional entry if you order plots with a column order}

\item{x}{jabba.rds}

\item{ref}{reference to use for pgvdb}

\item{type}{ppfit, do not change for this plot type}

\item{visible}{TRUE/FALSE whether the plot is hidden or showing in pgv}

\item{title}{title of the plot in pgvdb}

\item{annotation}{default is NULL, currently won't work for ppfit}

\item{overwrite}{TRUE/FALSE to overwrite an existing genome json}

\item{tag}{optional argument, can be binset to override y spacing in pgv}
}
\description{
function to create data.table for ppfit for pgvdb
}
\author{
Stanley Clarke
}
