% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\docType{data}
\name{priority_columns_jabba}
\alias{priority_columns_jabba}
\title{Jabba priority}
\format{
An object of class \code{character} of length 3.
}
\usage{
priority_columns_jabba
}
\description{
Non-allelic jabbas
}
\keyword{datasets}
