% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{process_cytoband}
\alias{process_cytoband}
\title{process_cytoband}
\usage{
process_cytoband(cyto = NULL, coarse = FALSE)
}
\arguments{
\item{cytoband}{path to cytoband file provided by UCSC}
}
\value{
GenomicRanges::GRanges object
}
\description{
Helper script to process cytoband
}
\author{
Kevin Hadi
}
