% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_jabba}
\alias{process_jabba}
\title{Read jabba}
\usage{
process_jabba(jabba, allow_return_null = FALSE)
}
\description{
Function to flexibly read in jabba files whether providing raw jabba list or ggraph
}
