% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{process_qc_flag}
\alias{process_qc_flag}
\title{QC Flags}
\usage{
process_qc_flag(metadata, qc_flag_thresholds)
}
\description{
Parse QC flags into strings
}
\details{
QC metrics from picard need to be parsed based on coverage, insert size
total number of reads, duplicate rate.
The strings should be parsable into a form
digested in gOS and shown as a single "PASS"/Checkmark", "Warning", or "Fail". 
The actual metrics should show up on hover.
}
