% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{read_decomposed_mutationtype_probabilities}
\alias{read_decomposed_mutationtype_probabilities}
\title{read_decomposed_mutationtype_probabilities}
\usage{
read_decomposed_mutationtype_probabilities(
  decomposed_probs_path,
  is_indel = FALSE
)
}
\arguments{
\item{decomposed_probs_path}{Path to sigprofiler decomposed probabilities file}

\item{is_indel}{Boolean to indicate if the input matrix is for indels}
}
\value{
data.frame containing formatted decomposed probabilities
}
\description{
Reads and formats decomposed probabilities from sigprofiler outputs for a single sample
}
\author{
Johnathan Rafailov, Shihab Dider
}
