% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-package.R
\docType{data}
\name{READ_LENGTH}
\alias{READ_LENGTH}
\title{Read length}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
READ_LENGTH
}
\description{
Multiply paired reads by this number for read length
}
\details{
To get binned coverage into read length level coverage,
multiply by read length (151)
}
\keyword{datasets}
