% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{read_meta_data_json}
\alias{read_meta_data_json}
\title{read_meta_data_json}
\usage{
read_meta_data_json(meta_json, patient_id)
}
\arguments{
\item{meta_json}{path to metadata.json for case reports}

\item{patient_id}{sample name for metadata_json}
}
\value{
data.table of the metadata.json for each sample
}
\description{
Reads in meta data jsons and gets signatures that are lists into the correct list format in the data.table. Used in create_distributions
}
\author{
Stanley Clarke
}
