% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{read_nf_log}
\alias{read_nf_log}
\title{Parse .nextflow.log}
\usage{
read_nf_log(nflog_path, max_lines = 1000000L)
}
\description{
Parse nextflow log files for launchdir and input options
to get samplesheet
}
\author{
Kevin Hadi
}
