% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift-wrappers.R
\docType{data}
\name{required_columns}
\alias{required_columns}
\title{Define required columns for each lifter}
\format{
An object of class \code{list} of length 16.
}
\usage{
required_columns
}
\description{
Define required columns for each lifter
}
\keyword{datasets}
