% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sage_count}
\alias{sage_count}
\title{sage_count}
\usage{
sage_count(vcf_path, genome)
}
\arguments{
\item{vcf_path}{Path to the SAGE VCF file}

\item{genome}{Reference genome name (e.g., "hg19", "hg38")}
}
\value{
data.table
}
\description{
takes in a SAGE vcf and returns a data.table with total count and count of variants with normal vaf greater than 0
}
\author{
Shihab Dider, Tanubrata Dey
}
