% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sage_qc}
\alias{sage_qc}
\title{sage_qc}
\usage{
sage_qc(vcf_path, genome, outfile, write_json = TRUE, return_table = TRUE)
}
\arguments{
\item{vcf_path}{Path to sage VCF file to be used (paired T-N/ Tumor only)}

\item{genome}{Reference genome name (e.g., "hg19", "hg38")}

\item{outfile}{Path to write JSON}

\item{write_json}{TRUE/FALSE whether to write the JSON}

\item{return_table}{TRUE/FALSE whether to return the data}
}
\description{
function to create json for sage qc plotting in case reports
}
\author{
Shihab Dider, Tanubrata Dey, Stanley Clarke
}
