% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{SAGEcounts}
\alias{SAGEcounts}
\title{SAGEcounts}
\usage{
SAGEcounts(
  vcf,
  seqnames_genome_width = c(1:22, "X", "Y"),
  type_return = "counts"
)
}
\arguments{
\item{vcf}{patient id to be added to pgvdb or case reports}

\item{seqnames_genome_width}{chromosomes to count variants in}

\item{type_return}{counts or dt, if counts returns counts for snv and snv count with a normal vaf greater than 0. If dt, returns the parsed vcf}
}
\value{
data.table
}
\description{
takes in a SAGE vcf and returns a data.table with total count and count of variants with normal vaf greater than 0
}
\author{
Tanubrata Dey, Stanley Clarke
}
