% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sbs_deconstructSigs2meta}
\alias{sbs_deconstructSigs2meta}
\title{sbs_deconstructSigs2meta}
\usage{
sbs_deconstructSigs2meta(sbs_file, sample)
}
\arguments{
\item{sbs_file}{path to rds outpyut from deconstructSigs}

\item{sample}{sample or pair name-needed to subset to the correct sample}
}
\description{
generates list of SBS signatures from the output of deconstruct sigs to add to metadata
}
\author{
Sukanya Panja, Stanley Clarke
}
