% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sbs_sigprofiler2meta}
\alias{sbs_sigprofiler2meta}
\title{sbs_sigprofiler2meta}
\usage{
sbs_sigprofiler2meta(sbs_file, sample)
}
\arguments{
\item{sbs_file}{path to Assignment_Solution_Activities.txt, output of sigprofiler}

\item{sample}{sample or pair name-needed to subset to the correct sample}
}
\value{
list of length 2, first being the signatures as a fraction and second being signatures as counts
}
\description{
generates list of SBS signature counts and percentages from the output of deconstruct sigs to add to metadata
}
\author{
Stanley Clarke
}
