% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-cell-annotate.R
\name{sc_annotation_config}
\alias{sc_annotation_config}
\title{sc_annotation_config}
\usage{
sc_annotation_config(...)
}
\arguments{
\item{...}{values to override}
}
\value{
named list
}
\description{
Paths and settings used by the single-cell annotation functions (nf-gos
checkout, container images, SnpEff cache, OncoKB gene list, hg38 GENCODE
files, OncoTree code). Pass named values to override the lab defaults.

The two hg38 GENCODE files are the hg38 versions of nf-gos's hg19 files:
cna_gencode is oncokb/gencode.v19.annotation.gtf.nochr.tsg.onco.rds (the same
399 OncoKB oncogenes / tumor suppressors on GENCODE v32), fusions_gencode is
fusions/hg19/gencode.v29lift37.annotation.nochr.5putr.expanded.rds rebuilt
on GENCODE v32 (transcripts with a 5' UTR get a 150 kb upstream exon "-1"
and UTR, as in the hg19 file).
}
\author{
Stanley Clarke
}
