% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shutup}
\alias{shutup}
\title{Silence outputs.}
\usage{
shutup(top_level_expr, capture_output_type = "message")
}
\description{
Capture error messages
but silence superfluous intermediate messages
printed from lift_all().
}
