% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sigprofiler_decomposed_probs_json}
\alias{sigprofiler_decomposed_probs_json}
\title{sigprofiler_decomposed_probs_json}
\usage{
sigprofiler_decomposed_probs_json(
  probs,
  is_indel = FALSE,
  data_dir,
  suffix = "_somatic",
  pairs = NULL,
  pair_name = NULL,
  cores = 1
)
}
\arguments{
\item{probs}{path to sigprofiler decomposed probabilities}

\item{is_indel}{Boolean to indicate if the input matrix is for indels}

\item{data_dir}{path to data directory of which each subdirectory contains jsons per patient}

\item{suffix}{expected suffix of sample name in decomposed_probs_json matrix, default = "_somatic" but have also seen "_0000"}

\item{pairs}{vector of samples for which to create jsons, by default = NULL, indicating process all patients}

\item{pair_name}{expected pair name; only supply if you know the pair name and the input table is only for one sample!}

\item{cores}{number of cores to parallelize with}
}
\value{
data.table or NULL
}
\description{
function to create decomposed probabilities json for case reports as adapted to sigprofiler outputs
}
\author{
Johnathan Rafailov
}
