% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skilift-package.R
\docType{data}
\name{snpeff_protein_coding_annotations}
\alias{snpeff_protein_coding_annotations}
\title{Snpeff Protein Coding Annotations}
\format{
An object of class \code{character} of length 20.
}
\usage{
snpeff_protein_coding_annotations
}
\description{
Nuff said.
}
\keyword{datasets}
