% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{strelka_qc}
\alias{strelka_qc}
\title{strelka_qc}
\usage{
strelka_qc(
  strelkaqc_filtered_rds,
  outfile,
  write_json = TRUE,
  return_table = TRUE
)
}
\arguments{
\item{strelkaqc_filtered_rds}{path to an rds of filtered strelka output}

\item{outfile}{path to write json}

\item{write_json}{TRUE/FALSE whether to write the json}

\item{return_table}{TRUE/FALSE whether to return the data}
}
\description{
function to create json for strelka qc plotting in case reports
}
\author{
Tanubrata Dey
}
