% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{subsample_hetsnps}
\alias{subsample_hetsnps}
\title{subsample_hetsnps}
\usage{
subsample_hetsnps(het_pileups, mask = NULL, sample_size = 1e+05)
}
\arguments{
\item{het_pileups}{sites.txt from het pileups}

\item{mask}{rds file with masked regions}

\item{sample_size}{number of snps to randomly sample}
}
\description{
subsets the hetsnps to masked unique sites and colors by major/minor allele
}
\author{
Jonathan Rafailov, Shihab Dider
}
