% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{test_paths}
\alias{test_paths}
\title{Test path vector}
\usage{
test_paths(
  objects,
  rds_regex = ".rds$",
  gff_regex = ".gtf(.gz){0,}$|.gff([0-9]){0,}(.gz){0,}$",
  bcf_regex = ".bcf(.bgz|.gz){0,}$",
  vcf_regex = ".vcf(.bgz|.gz){0,}$",
  verbose = TRUE
)
}
\description{
Test if paths exist robustly
}
