% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\name{unify_colmap_nfmap}
\alias{unify_colmap_nfmap}
\title{Create unified column and nf casereports map}
\usage{
unify_colmap_nfmap(col_mapping, path_patterns)
}
\description{
Parse column mapping and nf-casereport path mapping
}
\details{
Takes in a list of named column regex patterns and
a list of named nf-casereport path regex patterns and
coerces them into a unified object for later use.
To be added as an active binding.
Not exported.
}
