% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{vcf_count}
\alias{vcf_count}
\title{vcf_count}
\usage{
vcf_count(vcf_path, genome)
}
\arguments{
\item{vcf_path}{Path to the VCF file}

\item{genome}{Reference genome name (e.g., "hg19", "hg38")}
}
\value{
data.table
}
\description{
takes in a vcf and returns a data.table with total count and count of variants with normal vaf greater than 0
}
\author{
Shihab Dider, Tanubrata Dey
}
