% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{write_distributions_to_json}
\alias{write_distributions_to_json}
\title{write_distributions_to_json}
\usage{
write_distributions_to_json(
  distributions,
  common_dir,
  cores = 1,
  overwrite = FALSE,
  haveSignatures = TRUE
)
}
\arguments{
\item{distributions}{List of data tables containing distributions}

\item{common_dir}{path to the common directory in which to write the distribution JSONs}

\item{cores}{number of cores to use for parallel processing}

\item{overwrite}{boolean flag to indicate whether to overwrite existing files}
}
\description{
Writes the distributions data table to JSON files
}
