% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{write_signature_jsons}
\alias{write_signature_jsons}
\title{write_signature_jsons}
\usage{
write_signature_jsons(signatures_list, common_folder, cores = 1)
}
\arguments{
\item{signatures_list}{list outputted from convert_sign}

\item{common_folder}{location of the signatures folder}

\item{cores}{cores for generating the signature distribution files}
}
\description{
writes the different distribution jsons for signatures from the list outputted by convert_signature_meta_json
}
\author{
Stanley Clarke
}
