% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_coverage_metrics}
\alias{add_coverage_metrics}
\alias{add_conpair}
\title{Add Coverage Metrics}
\usage{
add_coverage_metrics(
  metadata,
  tumor_coverage = NULL,
  foreground_col_name = "foreground",
  estimate_library_complexity = NULL,
  alignment_summary_metrics = NULL,
  insert_size_metrics = NULL,
  tumor_wgs_metrics = NULL,
  normal_wgs_metrics = NULL
)

add_conpair(metadata, conpair_contamination = NULL, conpair_concordance = NULL)
}
\arguments{
\item{metadata}{A data.table containing metadata.}

\item{tumor_coverage}{Coverage data for the tumor.}

\item{foreground_col_name}{Name of the column in the coverage data to use for foreground coverage}

\item{estimate_library_complexity}{Path to library complexity metrics file}

\item{alignment_summary_metrics}{Path to alignment summary metrics file}

\item{insert_size_metrics}{Path to insert size metrics file}

\item{wgs_metrics}{Path to WGS metrics file}
}
\value{
Updated metadata with coverage metrics added.

Updated metadata with coverage metrics added.
}
\description{
Adds coverage metrics to the metadata based on tumor coverage and QC metrics.

Adds coverage metrics to the metadata based on tumor coverage and QC metrics.
}
