% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncotable.R
\name{collect_oncokb_cna}
\alias{collect_oncokb_cna}
\title{collect_oncokb_cna}
\usage{
collect_oncokb_cna(oncokb_cna, verbose = TRUE)
}
\arguments{
\item{oncokb_cna}{Path to the oncokb CNA file.}

\item{verbose}{Logical flag to indicate if messages should be printed.}
}
\value{
A data.table containing processed OncoKB CNA information.
}
\description{
Collects OncoKB CNA data from a specified file and processes it.
}
