% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_distributions}
\alias{create_distributions}
\title{create_distributions}
\usage{
create_distributions(
  case_reports_data_folder,
  common_folder,
  filter_patients = NULL,
  write_jsons = TRUE,
  make_sub_folders = TRUE,
  cores = 1
)
}
\arguments{
\item{case_reports_data_folder}{folder with all case report data by sample}

\item{common_folder}{path to a folder to write all 7 jsons}

\item{filter_pateints}{list of samples to filter on to create the distributions}
}
\description{
function to create the meta data summary for samples in a case reports instance
}
\author{
Stanley Clarke, Tanubrata Dey, Joel Rosiene
}
