% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_oncotable}
\alias{create_oncotable}
\title{create_oncotable}
\usage{
create_oncotable(
  cohort,
  amp_thresh_multiplier = 1.5,
  gencode = "~/DB/GENCODE/gencode.v29lift37.annotation.nochr.rds",
  outdir,
  cores = 1
)
}
\arguments{
\item{cohort}{data.table with columns: pair (required), annotated_bcf, signature_counts, fusions, jabba_simple, karyograph, events, oncokb_maf, oncokb_cna}

\item{amp_thresh_multiplier}{amp.thresh for oncotable is amp_thresh_multiplier*ploidy}

\item{gencode}{file to gencode annotations (uses v29lift37 by default)}

\item{outdir}{path to directory in which to write oncotable outputs}

\item{cores}{number of cores for parallel processing}
}
\value{
Named list of oncotable results
}
\description{
function to create oncotable for use with filtered_events_json
}
\author{
Shihab Dider, Joel Rosiene
}
