% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R, R/plots.R
\name{create_oncotable}
\alias{create_oncotable}
\title{create_oncotable}
\usage{
create_oncotable(
  pair,
  annotated_bcf,
  signature_counts = NULL,
  fusions,
  jabba_simple,
  karyograph,
  events,
  gencode = NULL,
  amp_thresh_multiplier = NULL,
  outdir
)

create_oncotable(
  pair,
  annotated_bcf,
  signature_counts = NULL,
  fusions,
  jabba_simple,
  karyograph,
  events,
  gencode = NULL,
  amp_thresh_multiplier = NULL,
  outdir
)
}
\arguments{
\item{pair}{patient id}

\item{annotated_bcf}{bcf with variant annotations (snpeff)}

\item{signature_counts}{output of Signatures task (optional)}

\item{fusions}{output of fusions task}

\item{jabba_simple}{jabba.simple.rds from JaBbA}

\item{events}{output of events task}

\item{gencode}{file to gencode annotations (uses v29lift37 by default)}

\item{amp_thresh_multiplier}{amp.thresh for oncotable is amp_thresh_multiplier*ploidy}

\item{outdir}{path to directory in which to write oncotable outputs}

\item{cohort}{Cohort object containing sample information}

\item{cores}{number of cores for parallel processing}
}
\value{
None

data.table or NULL
}
\description{
function to create oncotable for use with filtered_events_json

function to create oncotable for use with filtered_events_json
}
\author{
Shihab Dider, Joel Rosiene
}
