% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_ppfit_json}
\alias{create_ppfit_json}
\title{create_ppfit_json}
\usage{
create_ppfit_json(
  balanced_gg,
  cov_path = NULL,
  path_obj = NULL,
  settings_json,
  ref = NULL,
  max_na = NULL,
  out_file = NULL,
  write_json = TRUE,
  overwrite = FALSE,
  return_table = FALSE,
  cores = 1
)
}
\arguments{
\item{balanced_gg}{the balanced_gg ggraph or the path to it}

\item{path_obj}{path to flow directory, if null, must supply coverage path}

\item{settings_json}{path to settings.json (from PGV)}

\item{ref}{reference name}

\item{max_na}{max.na used with JaBbA}

\item{out_file}{location to write json}

\item{write_json}{TRUE/FALSE whether to write the json}

\item{overwrite}{TRUE/FALSE whether to overwrite existing file}

\item{return_table}{TRUE/FALSE whether to return the data}

\item{cores}{cores for JaBbA:::segstats}

\item{cov}{path to associated coverage, if null, will try to pull path from pairs table}
}
\value{
NULL or segstats table
}
\description{
function to create segmentation plots in case reports
}
\author{
Stanley Clarke, Tanubrata Dey
drcln_cov = readRDS(thisp$decomposed_cov[i])
}
