% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_ppfit_json}
\alias{create_ppfit_json}
\title{create_ppfit_json}
\usage{
create_ppfit_json(
  jabba_gg,
  path_obj,
  out_file = NULL,
  write_json = TRUE,
  overwrite = FALSE,
  return_table = FALSE,
  cores = 1
)
}
\arguments{
\item{out_file}{location to write json}

\item{write_json}{TRUE/FALSE whether to write the json}

\item{overwrite}{TRUE/FALSE whether to overwrite existing file}

\item{return_table}{TRUE/FALSE whether to return the data}

\item{cores}{cores for JaBbA:::segstats}

\item{jabba_rds}{list object jabba.rds}
}
\value{
NULL or segstats table
}
\description{
function to create segmentation plots in case reports
}
\author{
Stanley Clarke, Tanubrata Dey
drcln_cov = readRDS(thisp$decomposed_cov[i])
}
