% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_somatic_json}
\alias{create_somatic_json}
\title{create_somatic_json}
\usage{
create_somatic_json(
  somatic_snv_cn,
  out_file,
  pair,
  pgv_settings,
  return_table_pgv = FALSE,
  meta_keep = NULL,
  y_col = "est_cn_llrm"
)
}
\arguments{
\item{somatic_snv_cn}{somatic copy number data table}

\item{out_file}{where the json should be written}

\item{pair}{patient.id to be used for pgvdb or case reports}

\item{pgv_settings}{the settings file used for pgvdb or case reports}

\item{return_table_pgv}{TRUE/FALSE whether to return the data.table to add to pgvdb}

\item{meta_keep}{meta data to write in the json}

\item{y_col}{column to be used for y axis values}
}
\value{
data.table to add to pgv or NULL, depends on return_table_pgv
}
\description{
create_somatic_json
}
\author{
Stanley Clarke
}
