% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant-qc.R
\name{create_variant_qc}
\alias{create_variant_qc}
\title{create_variant_qc}
\usage{
create_variant_qc(
  oncokb_path,
  multiplicity_path,
  sbs_path,
  indel_path,
  cohort_type
)
}
\arguments{
\item{oncokb_path}{Path to oncokb maf file to be used}

\item{multiplicity_path}{Path to multiplitcity output}

\item{sbs_path}{Path to posterior probs for sbs signatures}

\item{indel_path}{Path to posterior probs for ID signatures}
}
\value{
data.table containing variant QC metrics along with highest signature posterior
}
\description{
Function to create quality control metrics for somatic variants
}
\author{
Kevin Hadi, Aditya Deshpande
}
