% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R, R/plots.R
\name{dlrs}
\alias{dlrs}
\title{derivative log ratio spread}
\usage{
dlrs(x)

dlrs(x)
}
\arguments{
\item{x}{foreground from dryclean}
}
\value{
dlrs

dlrs
}
\description{
function to get the dlrs for coverage data
used in meta_data_json

function to get the dlrs for coverage data
used in meta_data_json
}
\author{
Joel Rosiene
}
