% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{filtered_events_json}
\alias{filtered_events_json}
\title{filtered_events_json}
\usage{
filtered_events_json(
  pair,
  oncotable,
  jabba_gg,
  out_file,
  temp_fix = FALSE,
  return_table = FALSE
)
}
\arguments{
\item{pair}{patient id to be added to pgvdb or case reports}

\item{oncotable}{oncotable task output}

\item{jabba_gg}{JaBbA output ggraph or complex}

\item{out_file}{path to write json}

\item{return_table}{TRUE/FALSE whether to return the data.table that is used for creating the json}

\item{cgc_file}{path to cgc file to annotate drivers}

\item{oncokab_file}{path to oncokb gene tiers to annotate drivers}
}
\value{
data.table or NULL
}
\description{
function to create filtered events json for case reports
}
\author{
Stanley Clarke, Tanubrata Dey
}
