% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift-wrappers.R
\name{lift_all}
\alias{lift_all}
\title{Run all lift methods on a cohort}
\usage{
lift_all(
  cohort,
  output_data_dir,
  cores = 1,
  genome_length = NULL,
  width = 10000,
  gencode = Skilift::get_default_gencode(),
  ...
)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}

\item{...}{Additional arguments}
}
\value{
None
}
\description{
This function runs all available lift methods on a cohort to generate the complete
set of output files needed for visualization.
}
