% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-width-distribution.R
\name{lift_multiplicity_fits}
\alias{lift_multiplicity_fits}
\title{lift_multiplicity_fits}
\usage{
lift_multiplicity_fits(cohort, output_data_dir, cores = 1)
}
\arguments{
\item{cohort}{Cohort object containing sample information}

\item{output_data_dir}{Base directory for output files}

\item{cores}{Number of cores for parallel processing (default: 1)}
}
\description{
lift_multiplicity_fits
}
