% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cohort.R
\docType{data}
\name{nf_path_patterns}
\alias{nf_path_patterns}
\title{Expected nf-casereport patterns}
\format{
An object of class \code{list} of length 30.
}
\usage{
nf_path_patterns
}
\description{
Encoding the expected path variables
}
\details{
This variable contains the paths to be expected from
the nf-casereport directory structure. 
You can output the default template to modify it using
base::dput(Skilift::nf_path_patterns)
}
\keyword{datasets}
