% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{oncotable}
\alias{oncotable}
\title{oncotable}
\usage{
oncotable(
  pair,
  somatic_variant_annotations = NULL,
  fusions = NULL,
  jabba_gg = NULL,
  karyograph = NULL,
  events = NULL,
  signature_counts = NULL,
  oncokb_snv = NULL,
  oncokb_cna = NULL,
  oncokb_fusions = NULL,
  gencode,
  cytoband,
  verbose = TRUE,
  amp.thresh = 4,
  filter = "PASS",
  del.thresh = 0.5
)
}
\arguments{
\item{somatic_variant_annotations}{Path to annotated.bcf file}

\item{fusions}{Path to fusion.rds file}

\item{jabba_gg}{Path to jabba.simple.rds file}

\item{karyograph}{Optional path to the karyograph.rds file}

\item{events}{Path to complex.rds file}

\item{signature_counts}{Path to signature_counts.txt file}

\item{oncokb_snv}{Path to oncokb MAF file}

\item{oncokb_cna}{Path to oncokb CNA file}

\item{verbose}{logical flag}

\item{amp.thresh}{SCNA amplification threshold to call an amp as a function of ploidy (4)}

\item{del.thresh}{SCNA deletion threshold for (het) del as a function of ploidy (by default cn = 1 will be called del, but this allows additoinal regions in high ploidy tumors to be considered het dels)}

\item{gencode_gr}{GRanges object with gencode annotations}
}
\description{
oncotable
}
