% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered-events.R
\name{select_heme_events}
\alias{select_heme_events}
\title{Select Heme events from Addy's hemedb}
\usage{
select_heme_events(filtered_events, hemedb_path = Skilift:::HEMEDB())
}
\description{
Coarse selection of Heme events for first pass filtering
}
