% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{strelka2counts}
\alias{strelka2counts}
\title{strelka2counts}
\usage{
strelka2counts(vcf, seqnames_genome_width = c(1:22, "X", "Y"))
}
\arguments{
\item{vcf}{patient id to be added to pgvdb or case reports}

\item{seqnames_genome_width}{chromosomes to count variants in}
}
\value{
data.table
}
\description{
takes in a strelka vcf and returns a data.table with total count and count of variants with normal vaf greater than 0
}
\author{
Stanley Clarke, Tanubrata Dey
}
