% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{strelka_qc}
\alias{strelka_qc}
\title{strelka_qc}
\usage{
strelka_qc(
  vcf,
  seqnames_genome_width = c(1:22, "X", "Y"),
  outfile,
  write_json = TRUE,
  return_table = TRUE
)
}
\arguments{
\item{outfile}{path to write json}

\item{write_json}{TRUE/FALSE whether to write the json}

\item{return_table}{TRUE/FALSE whether to return the data}

\item{path}{to strelka2 vcf file to be used}

\item{chromosomes}{to select for. default: c(1:22,"X","Y")}
}
\description{
function to create json for strelka qc plotting in case reports
}
\author{
Tanubrata Dey
}
